test(channels): isolate reasoning truncation coverage
